+++ fleet-management-app/src/lib/getEquipmentData.js
