+++ fleet-management-app/src/lib/getPositionData.js
