+++ fleet-management-app/src/lib/getStateData.js
